Retry unmounts if the mount point is busy
A mount point in the sandbox can be left busy for a variety of reasons,
and in particular because of processes that haven't yet finished
executing after being killed.

We were already waiting for the file systems to disappear with a retry
loop, so extend the loop to cover the umount(8) calls too.