Make Xcode work within a darwin-native sandbox
bindfs-mounting Xcode makes it non-functional because the xcodebuild tool
relies on the getattrlist(2) system call, which neither OSXFUSE nor bindfs
implement.

As sandboxctl exists particularly to support building software in a
controlled environment, special-case this by introducing a
DARWIN_NATIVE_WITH_XCODE setting that causes Xcode to be replicated within
the sandbox.  This is not great but at least allows compilations to work.