Add rudimentary Linux support
This modifies the sandbox module internals to support Linux-specific
functionality and adds the linux_native sandbox type to create a
null-mounted sandbox on Linux.  This should be functional on most
Linux distributions but so far I have only tested this on a
Ubuntu 14.04 LTS.

Note that linux_native could also be applicable to other supported
operating systems, possibly all.  Renaming it to "generic" or similar
is a potential future improvement.