Mount all of /Library in the sandbox
Instead of whitelisting a small subset of directories from /Library
into the sandbox just so that we can create a read-write Caches, mount
everything and blacklist Caches.

This is necessary in case the system configuration refers to files not
previously mounted in /Library.  The latest case I hit was my system
configuration at work, where the PAM settings pointed at a special
module in a custom directory within /Library, and thus "su" calls failed
because of the missing module.  There is no way a whitelist could cover
that directory because the name was custom.