Be more selective about what we mount in /Library
The parent commit changed the darwin_native type to mount the majority
of the /Library subdirectories into the sandbox.  Unfortunately, OSXFUSE
doesn't like too many mount points: they are expensive memory-wise and,
depending on the OSXFUSE version, the limit of allowed mount points can
be quite log (e.g. 24).

To fix this, blacklist the set of /Library subdirectories that we know
are not very useful within the sandbox.  This allows any non-blacklisted
subdirectories to still be automatically mounted.