Warn if SIP is enabled when creating a macOS sandbox
SIP is becoming more and more hostile to what sandboxctl tries to do.
I haven't been able to diagnose yet why simply trying to run a binary
like clang from within bindfs causes the kernel to kill it immediately,
but that's the case.

This might be fixable or it might not be.  Diagnosing the cause of
these issues is very complicated.  For now, warn users if SIP is
enabled when creating a sandbox.

Also add a note about this to sandboxctl.conf(5).