Raise open files limit when running bindfs
Something has changed in either macOS High Sierra or Mojave that causes
xcodebuild to fail when running inside the sandbox, and the failure
indicates that bindfs ran out of file descriptors.  It is possible that
these macOS updates lowered the default maximum, or that some library
used by bindfs changed behavior.

To fix this, just increase the number of file descriptors bindfs can use
to the maximum allowed by the kernel.